TP2/requirements.txt
pyzmq==22.3.0
requests==2.25.0
